Conflicts:
	src/main/java/site/greenwave/board/BoardEntity.java
	src/main/java/site/greenwave/diary/DiaryEntity.java
	src/main/java/site/greenwave/file/FileUtil.java
	src/main/java/site/greenwave/member/MemberEntity.java
	src/main/java/site/greenwave/point/BillEntity.java
	src/main/java/site/greenwave/point/PointEntity.java

+++ src/main/java/site/greenwave/point/repository/PointRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 
 import site.greenwave.point.entity.PointEntity;
 
-public interface PointRepository extends JpaRepository<PointEntity, Long>, QuerydslPredicateExecutor<PointEntity>{
+public interface PointRepository extends JpaRepository<PointEntity, Integer>, QuerydslPredicateExecutor<PointEntity>{
 	
 	//포인트 내역: 현재 보유 포인트
 	@Query("SELECT " +
